+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "ledger-cli" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.1] - 2025-07-16
 
 - Initial release
+
+## [0.0.2] - 2025-07-17
+- Improved formatting
+- Added transaction status cycling command
+
+## [0.0.3] - 2025-09-17
+- Implemented account auto completion